* develop: (39 commits)
  Feature #188 restrict dates in forecast selector (#213)
  change precip accum color bar to viridis (#215)
  use evenly spaced font for logs
  fix bug showing actual value of email notify setting in job details view by adding notify to list of keys to pass
  fix bug getting nx/ny values that are incorrect when multiple domains are specified
  revert change to docs because we are not renaming logs zip file
  move exit button to top right for log viewer, override sort of keys for log file list, replace disabled button with span with set pixel width
  make Logs button blue and change Close button to an X icon on the top right of window
  added option to wrfcloud-run to skip deletion of cluster for running on by hand cluster
  fix bug in email notify logic by adding setting expected env var used to send email
  change logs zip file name back to exclude job ID since it can be added when the user downloads it
  improve formatting and display current log file as label of text area
  create job data directory to write wrfcloud log into it
  prevent error if user clicks on Loading Logs text before log list has been loaded
  write wrfcloud runtime log file into job directory so all files for a job are contained in the same directory and simplifies logic to get log files for log viewer
  add job ID to logs zip file so it is clear what it contains if/when we allow users to download it
  cleanup
  store name of log with directory to find file from tree selector
  store full name of log file so it can be easily referenced when viewing the file
  replace sidenav with tree for log viewer
  ...

# Conflicts:
#	web/src/app/app.module.ts

+++ python/src/wrfcloud/runtime/run.py
@@ -36,6 +36,7 @@ def main() -> None:
         log.debug('Reading command line arguments')
         parser = argparse.ArgumentParser()
         parser.add_argument('--job-id', type=str, help='Job ID with run details.', required=True)
+        parser.add_argument('--delete-cluster', action=argparse.BooleanOptionalAction, help='Delete cluster or not')
         args = parser.parse_args()
         job_id = args.job_id
 
@@ -125,7 +126,8 @@ def main() -> None:
     # Shutdown the cluster after completion or failure
     try:
         _save_log_files(job)
-        _delete_cluster()
+        if args.delete_cluster:
+            _delete_cluster()
     except Exception as e:
         log.error('Failed to delete cluster.', e)
         _update_job_status(job, WrfJob.STATUS_CODE_FAILED,
@@ -187,7 +189,7 @@ def _save_log_files(job: WrfJob) -> None:
     """
     # find the files
     log_files: List[str] = []
-    log_files += glob.glob(f'/data/wrfcloud-run-{job.job_id}.log')
+    log_files += glob.glob(f'{job.work_dir}/wrfcloud-run-{job.job_id}.log')
     log_files += glob.glob(f'{job.work_dir}/geogrid/geogrid.log')
     log_files += glob.glob(f'{job.work_dir}/ungrib/ungrib.log')
     log_files += glob.glob(f'{job.work_dir}/metgrid/metgrid.log')
